requirements.txt
browser-use>=1.0.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
pytest>=7.0.0
pytest-asyncio>=0.21.0